requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
networkx==3.3
orjson==3.10.7
rapidfuzz==3.9.7
rank-bm25==0.2.2
tqdm==4.66.5
httpx==0.27.2
streamlit==1.38.0
pandas==2.2.3
